Combine data_transfer.py to fallback.py
